+++ OpenBudgeteer.Core/Migrations/Sqlite/20220213195523_ImportCreditColumn.cs
@@ -1,0 +1,25 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace OpenBudgeteer.Core.Migrations.Sqlite
+{
+    public partial class ImportCreditColumn : Migration
+    {
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<string>(
+                name: "CreditColumnName",
+                table: "ImportProfile",
+                type: "TEXT",
+                nullable: true);
+        }
+
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "CreditColumnName",
+                table: "ImportProfile");
+        }
+    }
+}