+++ OpenBudgeteer.Core/ViewModels/ImportDataViewModel.cs
@@ -14,6 +14,7 @@ using TinyCsvParser;
 using TinyCsvParser.Mapping;
 using TinyCsvParser.Tokenizer.RFC4180;
 using TinyCsvParser.TypeConverter;
+using System.Linq.Expressions;
 
 namespace OpenBudgeteer.Core.ViewModels;
 
@@ -32,15 +33,68 @@ public class ImportDataViewModel : ViewModelBase
         /// <param name="identifiedColumns">Collection of all CSV columns</param>
         public CsvBankTransactionMapping(ImportProfile importProfile, IEnumerable<string> identifiedColumns) : base()
         {
-            // TODO Add User Input for CultureInfo for Amount & TransactionDate conversion
-
-            MapProperty(identifiedColumns.ToList().IndexOf(importProfile.AmountColumnName), x => x.Amount, new DecimalConverter(new CultureInfo(importProfile.NumberFormat)));
-            MapProperty(identifiedColumns.ToList().IndexOf(importProfile.MemoColumnName), x => x.Memo);
+            // Mandatory
+            MapProperty(
+                identifiedColumns.ToList().IndexOf(importProfile.MemoColumnName),
+                x => x.Memo);
+            MapProperty(
+                identifiedColumns.ToList().IndexOf(importProfile.TransactionDateColumnName),
+                x => x.TransactionDate, 
+                new DateTimeConverter(importProfile.DateFormat));
+            
+            // Optional
             if (!string.IsNullOrEmpty(importProfile.PayeeColumnName))
             {
-                MapProperty(identifiedColumns.ToList().IndexOf(importProfile.PayeeColumnName), x => x.Payee);
+                MapProperty(
+                    identifiedColumns.ToList().IndexOf(importProfile.PayeeColumnName), 
+                    x => x.Payee);
             }
-            MapProperty(identifiedColumns.ToList().IndexOf(importProfile.TransactionDateColumnName), x => x.TransactionDate, new DateTimeConverter(importProfile.DateFormat));
+
+            // Amount Mapping
+            if (!string.IsNullOrEmpty(importProfile.CreditColumnName))
+            {
+                MapUsing(((transaction, row) =>
+                {
+                    var debitValue = row.Tokens[identifiedColumns.ToList().IndexOf(importProfile.AmountColumnName)];
+                    var creditValue = row.Tokens[identifiedColumns.ToList().IndexOf(importProfile.CreditColumnName)];
+
+                    if (string.IsNullOrWhiteSpace(debitValue) && string.IsNullOrWhiteSpace(creditValue))
+                    {
+                        return false;
+                    }
+
+                    Decimal result;
+                    var converter = new DecimalConverter(new CultureInfo(importProfile.NumberFormat));
+
+                    if (string.IsNullOrWhiteSpace(debitValue))
+                    {
+                        var converterResult = converter.TryConvert(creditValue, out result);
+                        if (converterResult)
+                        {
+                            transaction.Amount = result > 0 ? result * -1 : result;
+                        }
+
+                        return converterResult;
+                    }
+                    else
+                    {
+                        var converterResult = converter.TryConvert(debitValue, out result);
+                        if (converterResult)
+                        {
+                            transaction.Amount = result;
+                        }
+                        
+                        return converterResult;
+                    }
+                }));
+            }
+            else
+            {
+                MapProperty(
+                    identifiedColumns.ToList().IndexOf(importProfile.AmountColumnName),
+                    x => x.Amount,
+                    new DecimalConverter(new CultureInfo(importProfile.NumberFormat)));
+            }
         }
     }
 
@@ -113,6 +167,16 @@ public class ImportDataViewModel : ViewModelBase
         get => _validRecords;
         private set => Set(ref _validRecords, value);
     }
+    
+    private int _potentialDuplicates;
+    /// <summary>
+    /// Number of records which have been identified as potential duplicate
+    /// </summary>
+    public int PotentialDuplicates
+    {
+        get => _potentialDuplicates;
+        private set => Set(ref _potentialDuplicates, value);
+    }
 
     private ObservableCollection<ImportProfile> _availableImportProfiles;
     /// <summary>
@@ -153,7 +217,17 @@ public class ImportDataViewModel : ViewModelBase
         get => _parsedRecords;
         private set => Set(ref _parsedRecords, value);
     }
-    
+
+    private ObservableCollection<Tuple<CsvMappingResult<BankTransaction>, List<BankTransaction>>> _duplicates;
+    /// <summary>
+    /// Collection of all parsed CSV records which are a potential duplicate of existing <see cref="BankTransaction"/> 
+    /// </summary>
+    public ObservableCollection<Tuple<CsvMappingResult<BankTransaction>, List<BankTransaction>>> Duplicates
+    {
+        get => _duplicates;
+        private set => Set(ref _duplicates, value);
+    }
+
     private bool _isProfileValid;
     private string[] _fileLines;
     private readonly DbContextOptions<DatabaseContext> _dbOptions;
@@ -168,6 +242,7 @@ public class ImportDataViewModel : ViewModelBase
         AvailableAccounts = new ObservableCollection<Account>();
         IdentifiedColumns = new ObservableCollection<string>();
         ParsedRecords = new ObservableCollection<CsvMappingResult<BankTransaction>>();
+        Duplicates = new ObservableCollection<Tuple<CsvMappingResult<BankTransaction>, List<BankTransaction>>>();
         SelectedImportProfile = new ImportProfile();
         SelectedAccount = new Account();
         _dbOptions = dbOptions;
@@ -272,7 +347,7 @@ public class ImportDataViewModel : ViewModelBase
     /// Loads all settings based on <see cref="SelectedImportProfile"/>
     /// </summary>
     /// <returns>Object which contains information and results of this method</returns>
-    public ViewModelOperationResult LoadProfile()
+    public async Task<ViewModelOperationResult> LoadProfileAsync()
     {
         try
         {
@@ -287,7 +362,7 @@ public class ImportDataViewModel : ViewModelBase
             var result = LoadHeaders();
             if (!result.IsSuccessful) throw new Exception(result.Message);
 
-            ValidateData();
+            await ValidateDataAsync();
             _isProfileValid = true;
 
             return new ViewModelOperationResult(true);
@@ -335,6 +410,8 @@ public class ImportDataViewModel : ViewModelBase
         TotalRecords = 0;
         RecordsWithErrors = 0;
         ValidRecords = 0;
+        PotentialDuplicates = 0;
+        Duplicates.Clear();
     }
 
     /// <summary>
@@ -345,7 +422,7 @@ public class ImportDataViewModel : ViewModelBase
     /// Sets also figures of the ViewModel like <see cref="TotalRecords"/> or <see cref="ValidRecords"/>
     /// </remarks>
     /// <returns>Object which contains information and results of this method</returns>
-    public ViewModelOperationResult ValidateData()
+    public async Task<ViewModelOperationResult> ValidateDataAsync()
     {
         try
         {
@@ -369,14 +446,14 @@ public class ImportDataViewModel : ViewModelBase
             var parsedResults = csvParser.ReadFromString(csvReaderOptions, stringBuilder.ToString()).ToList();
 
             ParsedRecords.Clear();
+            Duplicates.Clear();
             foreach (var parsedResult in parsedResults)
             {
                 ParsedRecords.Add(parsedResult);
             }
 
-            TotalRecords = parsedResults.Count;
-            RecordsWithErrors = parsedResults.Count(i => !i.IsValid);
-            ValidRecords = parsedResults.Count(i => i.IsValid);
+            await DuplicateCheckOnParsedRecordsAsync();
+            UpdateCountValues();
 
             if (ValidRecords > 0) _isProfileValid = true;
             return new ViewModelOperationResult(true);
@@ -386,23 +463,67 @@ public class ImportDataViewModel : ViewModelBase
             TotalRecords = 0;
             RecordsWithErrors = 0;
             ValidRecords = 0;
+            PotentialDuplicates = 0;
             ParsedRecords.Clear();
+            Duplicates.Clear();
             return new ViewModelOperationResult(false, e.Message);
         }
     }
 
     /// <summary>
+    /// Update counts of several statistics
+    /// </summary>
+    private void UpdateCountValues()
+    {
+        TotalRecords = ParsedRecords.Count;
+        RecordsWithErrors = ParsedRecords.Count(i => !i.IsValid);
+        ValidRecords = ParsedRecords.Count(i => i.IsValid);
+        PotentialDuplicates = Duplicates.Count;
+    }
+
+    /// <summary>
+    /// Checks each parsed CSV records on potential existing <see cref="BankTransaction"/> 
+    /// </summary>
+    private async Task DuplicateCheckOnParsedRecordsAsync()
+    {
+        foreach (var parsedRecord in ParsedRecords.Where(i => i.IsValid).ToList())
+        {
+            var duplicateSearches = Task.Run(() =>
+            {
+                using (var dbContext = new DatabaseContext(_dbOptions))
+                {
+                    var duplicates = dbContext.BankTransaction.Where(i =>
+                        i.TransactionDate.Date == parsedRecord.Result.TransactionDate.Date &&
+                        i.Amount == parsedRecord.Result.Amount &&
+                        (i.Memo == parsedRecord.Result.Memo || i.Payee == parsedRecord.Result.Payee));
+                    if (duplicates.Any())
+                        Duplicates.Add(new Tuple<CsvMappingResult<BankTransaction>, List<BankTransaction>>(parsedRecord, duplicates.ToList()));
+                }
+            });
+
+            await Task.WhenAll(duplicateSearches);
+        }
+    }
+
+    /// <summary>
+    /// Removes the passed duplicate from the parsed records to exclude it from import
+    /// </summary>
+    /// <param name="duplicate">Duplicate that should be excluded</param>
+    public void ExcludeDuplicateRecord(Tuple<CsvMappingResult<BankTransaction>, List<BankTransaction>> duplicate)
+    {
+        ParsedRecords.Remove(duplicate.Item1);
+        Duplicates.Remove(duplicate);
+        UpdateCountValues();
+    }
+
+    /// <summary>
     /// Uses data from <see cref="ParsedRecords"/> to store it in the database
     /// </summary>
-    /// <remarks>
-    /// This method will call <see cref="ValidateData"/>
-    /// </remarks>
     /// <returns>Object which contains information and results of this method</returns>
-    public ViewModelOperationResult ImportData()
+    public async Task<ViewModelOperationResult> ImportDataAsync()
     {
         if (!_isProfileValid) return new ViewModelOperationResult(false, "Unable to Import Data as current settings are invalid.");
         
-        ValidateData();
         using (var dbContext = new DatabaseContext(_dbOptions))
         {
             using (var transaction = dbContext.Database.BeginTransaction())